requirements.txt
PyYAML==6.0
inotify==0.2.10
pycryptodome==3.14.1